Remove MudPopoverService
