+++ src/MudBlazor.UnitTests/Extensions/ServiceCollectionExtensionsTests.cs
@@ -273,13 +273,9 @@ public class ServiceCollectionExtensionsTests
         // Act
         services.AddMudPopoverService();
         var serviceProvider = services.BuildServiceProvider();
-#pragma warning disable CS0618
-        var mudPopoverService = serviceProvider.GetService<IMudPopoverService>();
-#pragma warning restore CS0618
         var popoverService = serviceProvider.GetService<IPopoverService>();
 
         // Assert
-        mudPopoverService.Should().NotBeNull();
         popoverService.Should().NotBeNull();
     }
 
@@ -299,23 +295,15 @@ public class ServiceCollectionExtensionsTests
             options.ContainerClass = "container_class";
             options.FlipMargin = 100;
             options.ThrowOnDuplicateProvider = false;
-            options.Mode = PopoverMode.Legacy;
-#pragma warning disable CS0618 // Type or member is obsolete
-            options.PoolSize = 200;
-            options.PoolInitialFill = 10;
-#pragma warning restore CS0618 // Type or member is obsolete
+            options.Mode = PopoverMode.Default;
             expectedOptions = options;
         });
         var serviceProvider = services.BuildServiceProvider();
-#pragma warning disable CS0618
-        var mudPopoverService = serviceProvider.GetService<IMudPopoverService>();
-#pragma warning restore CS0618
         var popoverService = serviceProvider.GetService<IPopoverService>();
         var options = serviceProvider.GetRequiredService<IOptions<PopoverOptions>>();
         var actualOptions = options.Value;
 
         // Assert
-        mudPopoverService.Should().NotBeNull();
         popoverService.Should().NotBeNull();
         expectedOptions.Should().NotBeNull();
         actualOptions.Should().BeSameAs(expectedOptions);
@@ -431,9 +419,6 @@ public class ServiceCollectionExtensionsTests
         var jsEvent = serviceProvider.GetService<IJsEvent>();
         var jsEventFactory = serviceProvider.GetService<IJsEventFactory>();
         var scrollManager = serviceProvider.GetService<IScrollManager>();
-#pragma warning disable CS0618
-        var mudPopoverService = serviceProvider.GetService<IMudPopoverService>();
-#pragma warning restore CS0618
         var popoverService = serviceProvider.GetService<IPopoverService>();
         var scrollListener = serviceProvider.GetService<IScrollListener>();
         var scrollListenerFactory = serviceProvider.GetService<IScrollListenerFactory>();
@@ -455,7 +440,6 @@ public class ServiceCollectionExtensionsTests
         jsEvent.Should().NotBeNull();
         jsEventFactory.Should().NotBeNull();
         scrollManager.Should().NotBeNull();
-        mudPopoverService.Should().NotBeNull();
         popoverService.Should().NotBeNull();
         scrollListener.Should().NotBeNull();
         scrollListenerFactory.Should().NotBeNull();
@@ -515,11 +499,7 @@ public class ServiceCollectionExtensionsTests
             options.PopoverOptions.ContainerClass = "container_class";
             options.PopoverOptions.FlipMargin = 100;
             options.PopoverOptions.ThrowOnDuplicateProvider = false;
-            options.PopoverOptions.Mode = PopoverMode.Legacy;
-#pragma warning disable CS0618 // Type or member is obsolete
-            options.PopoverOptions.PoolSize = 300;
-            options.PopoverOptions.PoolInitialFill = 5;
-#pragma warning restore CS0618 // Type or member is obsolete
+            options.PopoverOptions.Mode = PopoverMode.Default;
 
             expectedOptions = options;
         });
@@ -533,9 +513,6 @@ public class ServiceCollectionExtensionsTests
         var jsEvent = serviceProvider.GetService<IJsEvent>();
         var jsEventFactory = serviceProvider.GetService<IJsEventFactory>();
         var scrollManager = serviceProvider.GetService<IScrollManager>();
-#pragma warning disable CS0618
-        var mudPopoverService = serviceProvider.GetService<IMudPopoverService>();
-#pragma warning restore CS0618
         var popoverService = serviceProvider.GetService<IPopoverService>();
         var scrollListener = serviceProvider.GetService<IScrollListener>();
         var scrollListenerFactory = serviceProvider.GetService<IScrollListenerFactory>();
@@ -565,7 +542,6 @@ public class ServiceCollectionExtensionsTests
         jsEvent.Should().NotBeNull();
         jsEventFactory.Should().NotBeNull();
         scrollManager.Should().NotBeNull();
-        mudPopoverService.Should().NotBeNull();
         popoverService.Should().NotBeNull();
         scrollListener.Should().NotBeNull();
         scrollListenerFactory.Should().NotBeNull();
@@ -583,10 +559,6 @@ public class ServiceCollectionExtensionsTests
         actualPopoverOptions.FlipMargin.Should().Be(expectedOptions.PopoverOptions.FlipMargin);
         actualPopoverOptions.ThrowOnDuplicateProvider.Should().Be(expectedOptions.PopoverOptions.ThrowOnDuplicateProvider);
         actualPopoverOptions.Mode.Should().Be(expectedOptions.PopoverOptions.Mode);
-#pragma warning disable CS0618 // Type or member is obsolete
-        actualPopoverOptions.PoolSize.Should().Be(expectedOptions.PopoverOptions.PoolSize);
-        actualPopoverOptions.PoolInitialFill.Should().Be(expectedOptions.PopoverOptions.PoolInitialFill);
-#pragma warning restore CS0618 // Type or member is obsolete
 
         actualResizeObserverOptions.EnableLogging.Should().Be(expectedOptions.ResizeObserverOptions.EnableLogging);
         actualResizeObserverOptions.ReportRate.Should().Be(expectedOptions.ResizeObserverOptions.ReportRate);
