+++ src/MudBlazor/Components/Popover/MudPopoverBase.cs
@@ -2,9 +2,7 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Diagnostics.CodeAnalysis;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
 using Microsoft.JSInterop;
 
@@ -20,17 +18,10 @@ namespace MudBlazor;
 /// </remarks>
 public abstract class MudPopoverBase : MudComponentBase, IPopover, IAsyncDisposable
 {
-    [Obsolete("For Legacy compatibility mode only, will be removed in a future version.")]
-    private MudPopoverHandler? _handler;
-
     private bool _afterFirstRender;
 
     /// <inheritdoc />
-    public virtual Guid Id { get; [Obsolete("Set is only needed for legacy mode only. Remove in a future version.")] private set; } = Guid.NewGuid();
-
-    [Inject]
-    [Obsolete($"Replaced by {nameof(PopoverService)}. Will be removed in a future version.")]
-    public IMudPopoverService Service { get; set; } = null!;
+    public virtual Guid Id { get; } = Guid.NewGuid();
 
     [Inject]
     protected IPopoverService PopoverService { get; set; } = null!;
@@ -58,15 +49,7 @@ public abstract class MudPopoverBase : MudComponentBase, IPopover, IAsyncDisposa
     /// <inheritdoc />
     protected override async Task OnInitializedAsync()
     {
-        if (PopoverService.PopoverOptions.Mode == PopoverMode.Legacy)
-#pragma warning disable CS0618 // Type or member is obsolete
-        {
-            _handler = Service.Register(ChildContent ?? new RenderFragment((x) => { }));
-            _handler.SetComponentBaseParameters(this, PopoverClass, PopoverStyles, Open);
-            Id = _handler.Id;
-        }
-#pragma warning restore CS0618 // Type or member is obsolete
-        else
+        if (PopoverService.PopoverOptions.Mode == PopoverMode.Default)
         {
             await PopoverService.CreatePopoverAsync(this);
         }
@@ -81,16 +64,7 @@ public abstract class MudPopoverBase : MudComponentBase, IPopover, IAsyncDisposa
 
         if (_afterFirstRender)
         {
-            if (PopoverService.PopoverOptions.Mode == PopoverMode.Legacy)
-#pragma warning disable CS0618 // Type or member is obsolete
-            {
-                if (_handler is not null)
-                {
-                    await _handler.UpdateFragmentAsync(ChildContent, this, PopoverClass, PopoverStyles, Open);
-                }
-            }
-#pragma warning restore CS0618 // Type or member is obsolete
-            else
+            if (PopoverService.PopoverOptions.Mode == PopoverMode.Default)
             {
                 await PopoverService.UpdatePopoverAsync(this);
             }
@@ -102,18 +76,7 @@ public abstract class MudPopoverBase : MudComponentBase, IPopover, IAsyncDisposa
     {
         if (firstRender)
         {
-            if (PopoverService.PopoverOptions.Mode == PopoverMode.Legacy)
-#pragma warning disable CS0618 // Type or member is obsolete
-            {
-                if (_handler is not null)
-                {
-                    await _handler.Initialize();
-                    await Service.InitializeIfNeeded();
-                    await _handler.UpdateFragmentAsync(ChildContent, this, PopoverClass, PopoverStyles, Open);
-                }
-            }
-#pragma warning restore CS0618 // Type or member is obsolete
-            else
+            if (PopoverService.PopoverOptions.Mode == PopoverMode.Default)
             {
                 await PopoverService.UpdatePopoverAsync(this);
             }
@@ -132,13 +95,7 @@ public abstract class MudPopoverBase : MudComponentBase, IPopover, IAsyncDisposa
         {
             if (IsJSRuntimeAvailable)
             {
-#pragma warning disable CS0618 // Type or member is obsolete
-                if (PopoverService.PopoverOptions.Mode == PopoverMode.Legacy)
-                {
-                    await Service.Unregister(_handler);
-                }
-#pragma warning restore CS0618 // Type or member is obsolete
-                else
+                if (PopoverService.PopoverOptions.Mode == PopoverMode.Default)
                 {
                     await PopoverService.DestroyPopoverAsync(this);
                 }
