+++ src/MudBlazor/Components/Popover/MudPopoverProvider.razor
@@ -3,8 +3,8 @@
 @if (Enabled)
 {
     <div class="mud-popover-provider">
-	    @foreach (var handler in GetActivePopovers())
-	    {
+        @foreach (var handler in PopoverService.ActivePopovers)
+        {
             <MudRender @ref="handler.ElementReference" @key="handler.Id">
 		        <div id="@($"popovercontent-{handler.Id}")" data-ticks="@(handler.ActivationDate?.Ticks ?? 0)" @attributes="@handler.UserAttributes" class="@handler.Class" style="@handler.Style">
 			        @if (handler.ShowContent)
