+++ src/MudBlazor/Components/Popover/MudPopoverProvider.razor.cs
@@ -2,11 +2,6 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
 using MudBlazor.Interfaces;
 
@@ -18,14 +13,10 @@ namespace MudBlazor
         private bool _isConnectedToService = false;
 
         [Inject]
-        [Obsolete($"Use {nameof(PopoverService)} instead. This will be removed in a future version")]
-        public IMudPopoverService Service { get; set; } = null!;
-
-        [Inject]
         internal IPopoverService PopoverService { get; set; } = null!;
 
         /// <summary>
-        /// In some scenarios we need more than one ThemeProvider but we must not have more than one
+        /// In some scenarios we need more than one ThemeProvider, but we must not have more than one
         /// PopoverProvider. Set a cascading value with UsePopoverProvider=false to prevent it.
         /// </summary>
         [CascadingParameter(Name = "UsePopoverProvider")]
@@ -33,10 +24,6 @@ namespace MudBlazor
 
         public void Dispose()
         {
-#pragma warning disable CS0618
-            //TODO: For backward compatibility with old service. Should be removed in a future version
-            Service.FragmentsChanged -= Service_FragmentsChanged;
-#pragma warning restore CS0618
             PopoverService.Unsubscribe(this);
         }
 
@@ -47,10 +34,6 @@ namespace MudBlazor
                 return;
             }
 
-#pragma warning disable CS0618
-            //TODO: For backward compatibility with old service. Should be removed in a future version
-            Service.FragmentsChanged += Service_FragmentsChanged;
-#pragma warning restore CS0618
             PopoverService.Subscribe(this);
             _isConnectedToService = true;
         }
@@ -61,25 +44,11 @@ namespace MudBlazor
 
             if (!Enabled && _isConnectedToService)
             {
-#pragma warning disable CS0618
-                //TODO: For backward compatibility with old service. Should be removed in a future version with the _isConnectedToService
-                Service.FragmentsChanged -= Service_FragmentsChanged;
-#pragma warning restore CS0618
                 PopoverService.Unsubscribe(this);
                 _isConnectedToService = false;
 
                 return;
             }
-
-#pragma warning disable CS0618
-            //TODO: For backward compatibility with old service. Whole block should be removed in a future version
-            if (Enabled && !_isConnectedToService)
-            {
-                Service.FragmentsChanged -= Service_FragmentsChanged; // make sure to avoid multiple registration
-                Service.FragmentsChanged += Service_FragmentsChanged;
-                _isConnectedToService = true;
-            }
-#pragma warning restore CS0618
 
             // Let's in our new case ignore _isConnectedToService and always update the subscription except Enabled = false. The manager is specifically designed for it.
             // The reason is because If an observer throws an exception during the PopoverCollectionUpdatedNotification, indicating a malfunction, it will be automatically unsubscribed.
@@ -87,7 +56,6 @@ namespace MudBlazor
             {
                 PopoverService.Subscribe(this);
             }
-
         }
 
         protected override async Task OnAfterRenderAsync(bool firstRender)
@@ -101,20 +69,6 @@ namespace MudBlazor
             }
             await base.OnAfterRenderAsync(firstRender);
         }
-
-        //TODO: For backward compatibility with old service. Should be removed in a future version
-        private void Service_FragmentsChanged(object? sender, EventArgs e)
-        {
-            InvokeAsync(StateHasChanged);
-        }
-
-        //TODO: For backward compatibility with old service. Should be removed in a future version
-#pragma warning disable CS0618
-        private IEnumerable<IMudPopoverHolder> GetActivePopovers()
-        {
-            return PopoverService.ActivePopovers.Concat(Service.Handlers);
-        }
-#pragma warning restore CS0618
 
         /// <inheritdoc />
         Guid IPopoverObserver.Id { get; } = Guid.NewGuid();
