+++ src/MudBlazor/Services/Popover/IPopoverObserver.cs
@@ -2,10 +2,6 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-
 namespace MudBlazor;
 
 #nullable enable
