+++ src/MudBlazor/Services/Popover/MudPopoverHolder.cs
@@ -2,8 +2,6 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
 using Microsoft.AspNetCore.Components;
 
 namespace MudBlazor;
