+++ src/MudBlazor/Services/Popover/PopoverHolderContainer.cs
@@ -2,8 +2,6 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections.Generic;
-
 namespace MudBlazor;
 
 #nullable enable
